Setup Basic authentication using Google Sign In. This implementation currently does not connect to a database yet and is simply to test authentication using JWT tokens and verification using google java api libraries.
The goal here is to be able to have the frontend send the backend an id_token received from authenticating using Google Oauth. This id_token is then verified using the Google Java Api Library which also gives us the user info we request such as name and email. This information can be used to generate a user entry in our database and authenticate against. If the id_token is valid and an account is made (or an existing account is found) then a JWT token is sent back to the front-end to use for authenticating API requests.